unified tcp and udp collecting process
