+++ pkg/collector/process.go
@@ -27,7 +27,6 @@ import (
 
 	"github.com/vmware/go-ipfix/pkg/config"
 	"github.com/vmware/go-ipfix/pkg/entities"
-	"github.com/vmware/go-ipfix/pkg/registry"
 )
 
 type collectingProcess struct {
@@ -37,9 +36,6 @@ type collectingProcess struct {
 	templatesLock *sync.RWMutex
 	// template lifetime
 	templateTTL uint32
-	// registries for decoding Information Element
-	ianaRegistry   registry.Registry
-	antreaRegistry registry.Registry
 	// server information
 	address net.Addr
 	// maximum buffer size to read the record
@@ -54,6 +50,120 @@ type templateField struct {
 	elementID     uint16
 	elementLength uint16
 	enterpriseID  uint32
+}
+
+func InitCollectingProcess(address net.Addr, maxBufferSize uint16, templateTTL uint32) (*collectingProcess, error) {
+	collectProc := &collectingProcess{
+		templatesMap:   make(map[uint32]map[uint16][]*templateField),
+		templatesLock:  &sync.RWMutex{},
+		templateTTL:    templateTTL,
+		address:        address,
+		maxBufferSize:  maxBufferSize,
+		stopChan:       make(chan bool),
+		messages:       make([]*entities.Message, 0),
+	}
+	return collectProc, nil
+}
+
+func (cp *collectingProcess) Start() {
+	if cp.address.Network() == "tcp" {
+		cp.startTCPServer()
+	} else if cp.address.Network() == "udp" {
+		cp.startUDPServer()
+	}
+}
+
+func (cp *collectingProcess) startTCPServer() {
+	listener, err := net.Listen("tcp", cp.address.String())
+	if err != nil {
+		klog.Errorf("Cannot start collecting process on %s: %v", cp.address.String(), err)
+		return
+	}
+
+	klog.Infof("Start %s collecting process on %s", cp.address.Network(), cp.address.String())
+	defer listener.Close()
+
+	for {
+		conn, err := listener.Accept()
+		if err != nil {
+			klog.Errorf("Cannot start collecting process on %s: %v", cp.address.String(), err)
+			return
+		}
+		errChan := make(chan bool)
+		go func() {
+			cp.handleConnection(conn, cp.maxBufferSize, cp.address, errChan)
+		}()
+		select {
+		case <-errChan:
+			return
+		}
+	}
+}
+
+func (cp *collectingProcess) startUDPServer() {
+	s, err := net.ResolveUDPAddr("udp", cp.address.String())
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	conn, err := net.ListenUDP("udp", s)
+	if err != nil {
+		fmt.Println(err)
+		return
+	}
+	defer conn.Close()
+
+	for {
+		errChan := make(chan bool)
+		go func() {
+			cp.handleConnection(conn, cp.maxBufferSize, cp.address, errChan)
+		}()
+		select {
+		case <-errChan:
+			return
+		}
+	}
+}
+
+func (cp *collectingProcess) handleConnection(conn net.Conn, maxBufferSize uint16, address net.Addr, errChan chan bool) {
+	for {
+		select {
+		case <-cp.stopChan:
+			errChan <- true
+			return
+		default:
+			buff := make([]byte, maxBufferSize)
+			size, err := conn.Read(buff)
+			if err != nil {
+				if err == io.EOF {
+					klog.Infof("Connection %s has been closed.", address.String())
+				} else {
+					klog.Errorf("Error in collecting process: %v", err)
+				}
+				return
+			}
+			klog.Infof("Receiving %d bytes from %s", size, address.String())
+			message := make([]byte, size)
+			copy(message, buff[0:size])
+			for size > 0 {
+				length, err := getMessageLength(bytes.NewBuffer(message))
+				if err != nil {
+					klog.Error(err)
+					return
+				}
+				size = size - length
+				// get the packet here
+				packet, err := cp.decodePacket(bytes.NewBuffer(message[0:length]))
+				if err != nil {
+					klog.Error(err)
+					return
+				}
+				klog.Info(packet)
+				message = message[length:]
+			}
+		}
+	}
+	conn.Close()
 }
 
 func (cp *collectingProcess) decodePacket(packetBuffer *bytes.Buffer) (*entities.Message, error) {
@@ -188,4 +298,15 @@ func decode(buffer io.Reader, output ...interface{}) error {
 		err = binary.Read(buffer, binary.BigEndian, out)
 	}
 	return err
+}
+
+// get buffer length by decoding the header
+func getMessageLength(msgBuffer *bytes.Buffer) (int, error) {
+	packet := entities.Message{}
+	var id, length uint16
+	err := decode(msgBuffer, &packet.Version, &packet.BufferLength, &packet.ExportTime, &packet.SeqNumber, &packet.ObsDomainID, &id, &length)
+	if err != nil {
+		return 0, fmt.Errorf("Cannot decode message: %v", err)
+	}
+	return int(packet.BufferLength), nil
 }
